+++ Robust.Shared/GameObjects/Systems/SharedUserInterfaceSystem.cs
@@ -1,13 +1,29 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
+using System.Linq;
+using System.Runtime.InteropServices;
+using JetBrains.Annotations;
 using Robust.Shared.Enums;
+using Robust.Shared.GameObjects;
 using Robust.Shared.Player;
+using Robust.Shared.Utility;
 
 namespace Robust.Shared.GameObjects;
 
 public abstract class SharedUserInterfaceSystem : EntitySystem
 {
+    /*
+     * TODO:
+     * Need the external call methods that raise the event as a predicted event(?)
+     * Need to handle closing via event
+     * Opening gets handled directly
+     * Need to be able to call open in a shared context.
+     * When changing mob need to close old UIs and open new ones (internally, don't call the event?)
+     * All events get raised shared maybe? Like uhh open UI or close UI or interact with it
+     * Server messages only get sent to relevant client.
+     */
+
     private EntityQuery<UserInterfaceComponent> _uiQuery;
 
     public override void Initialize()
@@ -50,6 +66,118 @@ public abstract class SharedUserInterfaceSystem : EntitySystem
         // Needs to close UIs that aren't networked anymore.
 
         throw new NotImplementedException();
+    }
+
+    /// <summary>
+    /// Closes the attached UI for all entities.
+    /// </summary>
+    public void CloseUi(Entity<UserInterfaceComponent?> entity, Enum key)
+    {
+        if (!_uiQuery.Resolve(entity.Owner, ref entity.Comp, false))
+            return;
+
+        // TODO: Need an event raised
+        foreach ()
+    }
+
+    /// <summary>
+    /// Closes the attached UI only for the specified actor.
+    /// </summary>
+    public void CloseUi(Entity<UserInterfaceComponent?> entity, Enum key, EntityUid actor)
+    {
+        if (!_uiQuery.Resolve(entity.Owner, ref entity.Comp, false))
+            return;
+    }
+
+    /// <summary>
+    /// Closes the attached UI only for the specified actor.
+    /// </summary>
+    public void CloseUi(Entity<UserInterfaceComponent?> entity, Enum key, ICommonSession actor)
+    {
+        var actorEnt = actor.AttachedEntity;
+
+        if (actorEnt == null)
+            return;
+
+        CloseUi(entity, key, actorEnt.Value);
+    }
+
+    public void OpenUi(Entity<UserInterfaceComponent?> entity, Enum key, EntityUid actor)
+    {
+
+    }
+
+    public void OpenUi(Entity<UserInterfaceComponent?> entity, Enum key, ICommonSession actor)
+    {
+        var actorEnt = actor.AttachedEntity;
+
+        if (actorEnt == null)
+            return;
+
+        OpenUi(entity, key, actorEnt.Value);
+    }
+
+    public void SetUiState(Entity<UserInterfaceComponent?> entity, Enum key, BoundUserInterfaceState? state)
+    {
+        if (!_uiQuery.Resolve(entity.Owner, ref entity.Comp, false))
+            return;
+
+        if (!entity.Comp.Interfaces.ContainsKey(key))
+            return;
+
+        ref var stateRef = ref CollectionsMarshal.GetValueRefOrAddDefault(entity.Comp.States, key, out var exists);
+
+        if (exists && state?.Equals(stateRef) == true)
+            return;
+
+        stateRef = state;
+        Dirty(entity);
+    }
+
+    public void SendUiMessage(Entity<UserInterfaceComponent?> entity, Enum key, BoundUserInterfaceMessage message)
+    {
+        if (!_uiQuery.Resolve(entity.Owner, ref entity.Comp, false))
+            return;
+
+        if (!entity.Comp.Actors.TryGetValue(key, out var actors))
+            return;
+
+        var filter = Filter.Entities(actors.ToArray());
+        RaiseNetworkEvent(new BoundUIWrapMessage(GetNetEntity(entity.Owner), message, key), filter);
+    }
+
+    public void SendUiMessage(Entity<UserInterfaceComponent?> entity, Enum key, BoundUserInterfaceMessage message, EntityUid actor)
+    {
+        if (!_uiQuery.Resolve(entity.Owner, ref entity.Comp, false))
+            return;
+
+        if (!entity.Comp.Actors.TryGetValue(key, out var actors) || !actors.Contains(actor))
+            return;
+
+        RaiseNetworkEvent(new BoundUIWrapMessage(GetNetEntity(entity.Owner), message, key), actor);
+    }
+
+    public void SendUiMessage(Entity<UserInterfaceComponent?> entity, Enum key, BoundUserInterfaceMessage message, ICommonSession actor)
+    {
+        if (!_uiQuery.Resolve(entity.Owner, ref entity.Comp, false) || actor.AttachedEntity is not { } attachedEntity)
+            return;
+
+        if (!entity.Comp.Actors.TryGetValue(key, out var actors) || !actors.Contains(attachedEntity))
+            return;
+
+        RaiseNetworkEvent(new BoundUIWrapMessage(GetNetEntity(entity.Owner), message, key), actor);
+    }
+
+    /// <summary>
+    /// Closes all UIs for the entity.
+    /// </summary>
+    public void CloseUis(Entity<UserInterfaceComponent?> entity)
+    {
+        if (!_uiQuery.Resolve(entity.Owner, ref entity.Comp, false))
+            return;
+
+        entity.Comp.Interfaces.Clear();
+        Dirty(entity);
     }
 
     /// <summary>
@@ -115,23 +243,6 @@ public abstract class SharedUserInterfaceSystem : EntitySystem
     }
 
     /// <summary>
-    /// Tries to get the BUIs if it's castable to the specified type.
-    /// </summary>
-    public IEnumerable<T> GetUis<T>(EntityUid uid, UserInterfaceComponent? ui = null)
-    {
-        if (!_uiQuery.Resolve(uid, ref ui, false))
-            yield break;
-
-        foreach (var bui in ui.Interfaces.Values)
-        {
-            if (bui is not T cast)
-                continue;
-
-            yield return cast;
-        }
-    }
-
-    /// <summary>
     /// Tries to get the BUI if it is currently open.
     /// </summary>
     public bool TryGetOpenUi(EntityUid uid, Enum uiKey, [NotNullWhen(true)] out BoundUserInterface? bui, UserInterfaceComponent? ui = null)
@@ -139,23 +250,6 @@ public abstract class SharedUserInterfaceSystem : EntitySystem
         bui = null;
 
         return _uiQuery.Resolve(uid, ref ui, false) && ui.ClientOpenInterfaces.TryGetValue(uiKey, out bui);
-    }
-
-    /// <summary>
-    /// Tries to get the open BUIs if it's castable to the specified type.
-    /// </summary>
-    public IEnumerable<T> GetOpenUis<T>(EntityUid uid, UserInterfaceComponent? ui = null)
-    {
-        if (!_uiQuery.Resolve(uid, ref ui, false))
-            yield break;
-
-        foreach (var bui in ui.ClientOpenInterfaces.Values)
-        {
-            if (bui is not T cast)
-                continue;
-
-            yield return cast;
-        }
     }
 
     /// <summary>
@@ -261,23 +355,6 @@ public abstract class SharedUserInterfaceSystem : EntitySystem
         return true;
     }
 
-    /*
-     * TODO:
-     * Need the external call methods that raise the event as a predicted event(?)
-     * Need to handle closing via event
-     * Opening gets handled directly
-     * Need to be able to call open in a shared context.
-     * When changing mob need to close old UIs and open new ones (internally, don't call the event?)
-     */
-
-    /// <summary>
-    /// Raised by client-side UIs to send to server.
-    /// </summary>
-    internal void SendUiMessage(BoundUserInterface bui, BoundUserInterfaceMessage msg)
-    {
-        RaiseNetworkEvent(new BoundUIWrapMessage(GetNetEntity(bui.Owner), msg, bui.UiKey));
-    }
-
     /// <summary>
     /// Raised by client-side UIs to send predicted messages to server.
     /// </summary>
@@ -285,4 +362,252 @@ public abstract class SharedUserInterfaceSystem : EntitySystem
     {
         RaisePredictiveEvent(new PredictedBoundUIWrapMessage(GetNetEntity(bui.Owner), msg, bui.UiKey));
     }
+
+    /// <inheritdoc />
+        public override void Update(float frameTime)
+        {
+            var xformQuery = GetEntityQuery<TransformComponent>();
+            var query = AllEntityQuery<ActiveUserInterfaceComponent, TransformComponent>();
+
+            while (query.MoveNext(out var uid, out var activeUis, out var xform))
+            {
+                foreach (var ui in activeUis.Interfaces)
+                {
+                    CheckRange(uid, activeUis, ui, xform, xformQuery);
+
+                    if (!ui.StateDirty)
+                        continue;
+
+                    ui.StateDirty = false;
+
+                    foreach (var (player, state) in ui.PlayerStateOverrides)
+                    {
+                        RaiseNetworkEvent(state, player.Channel);
+                    }
+
+                    if (ui.LastStateMsg == null)
+                        continue;
+
+                    foreach (var session in ui.SubscribedSessions)
+                    {
+                        if (!ui.PlayerStateOverrides.ContainsKey(session))
+                            RaiseNetworkEvent(ui.LastStateMsg, session.Channel);
+                    }
+                }
+            }
+        }
+
+        /// <summary>
+        ///     Verify that the subscribed clients are still in range of the interface.
+        /// </summary>
+        private void CheckRange(EntityUid uid, ActiveUserInterfaceComponent activeUis, PlayerBoundUserInterface ui, TransformComponent transform, EntityQuery<TransformComponent> query)
+        {
+            if (ui.InteractionRange <= 0)
+                return;
+
+            // We have to cache the set of sessions because Unsubscribe modifies the original.
+            _sessionCache.Clear();
+            _sessionCache.AddRange(ui.SubscribedSessions);
+
+            var uiPos = _xformSys.GetWorldPosition(transform, query);
+            var uiMap = transform.MapID;
+
+            foreach (var session in _sessionCache)
+            {
+                // The component manages the set of sessions, so this invalid session should be removed soon.
+                if (!query.TryGetComponent(session.AttachedEntity, out var xform))
+                    continue;
+
+                if (_ignoreUIRangeQuery.HasComponent(session.AttachedEntity))
+                    continue;
+
+                // Handle pluggable BoundUserInterfaceCheckRangeEvent
+                var checkRangeEvent = new BoundUserInterfaceCheckRangeEvent(uid, ui, session);
+                RaiseLocalEvent(uid, ref checkRangeEvent, broadcast: true);
+                if (checkRangeEvent.Result == BoundUserInterfaceRangeResult.Pass)
+                    continue;
+
+                if (checkRangeEvent.Result == BoundUserInterfaceRangeResult.Fail)
+                {
+                    CloseUi(ui, session, activeUis);
+                    continue;
+                }
+
+                DebugTools.Assert(checkRangeEvent.Result == BoundUserInterfaceRangeResult.Default);
+
+                if (uiMap != xform.MapID)
+                {
+                    CloseUi(ui, session, activeUis);
+                    continue;
+                }
+
+                var distanceSquared = (uiPos - _xformSys.GetWorldPosition(xform, query)).LengthSquared();
+                if (distanceSquared > ui.InteractionRangeSqrd)
+                    CloseUi(ui, session, activeUis);
+            }
+        }
+
+        #region Get BUI
+
+        public bool HasUi(EntityUid uid, Enum uiKey, UserInterfaceComponent? ui = null)
+        {
+            if (!Resolve(uid, ref ui))
+                return false;
+
+            return ui.Interfaces.ContainsKey(uiKey);
+        }
+
+        public PlayerBoundUserInterface GetUi(EntityUid uid, Enum uiKey, UserInterfaceComponent? ui = null)
+        {
+            if (!Resolve(uid, ref ui))
+                throw new InvalidOperationException($"Cannot get {typeof(PlayerBoundUserInterface)} from an entity without {typeof(UserInterfaceComponent)}!");
+
+            return ui.Interfaces[uiKey];
+        }
+
+        public PlayerBoundUserInterface? GetUiOrNull(EntityUid uid, Enum uiKey, UserInterfaceComponent? ui = null)
+        {
+            return TryGetUi(uid, uiKey, out var bui, ui)
+                ? bui
+                : null;
+        }
+
+        /// <summary>
+        ///     Return UIs a session has open.
+        ///     Null if empty.
+        /// </summary>
+        public List<PlayerBoundUserInterface>? GetAllUIsForSession(ICommonSession session)
+        {
+            OpenInterfaces.TryGetValue(session, out var value);
+            return value;
+        }
+        #endregion
+
+        public bool IsUiOpen(EntityUid uid, Enum uiKey, UserInterfaceComponent? ui = null)
+        {
+            if (!TryGetUi(uid, uiKey, out var bui, ui))
+                return false;
+
+            return bui.SubscribedSessions.Count > 0;
+        }
+
+        public bool SessionHasOpenUi(EntityUid uid, Enum uiKey, ICommonSession session, UserInterfaceComponent? ui = null)
+        {
+            if (!TryGetUi(uid, uiKey, out var bui, ui))
+                return false;
+
+            return bui.SubscribedSessions.Contains(session);
+        }
+
+        #region SendMessage
+
+        /// <summary>
+        ///     Send a BUI message to all connected player sessions.
+        /// </summary>
+        public bool TrySendUiMessage(EntityUid uid, Enum uiKey, BoundUserInterfaceMessage message, UserInterfaceComponent? ui = null)
+        {
+            if (!TryGetUi(uid, uiKey, out var bui, ui))
+                return false;
+
+            SendUiMessage(bui, message);
+            return true;
+        }
+
+        /// <summary>
+        ///     Send a BUI message to all connected player sessions.
+        /// </summary>
+        public void SendUiMessage(PlayerBoundUserInterface bui, BoundUserInterfaceMessage message)
+        {
+            var msg = new BoundUIWrapMessage(GetNetEntity(bui.Owner), message, bui.UiKey);
+            foreach (var session in bui.SubscribedSessions)
+            {
+                RaiseNetworkEvent(msg, session.Channel);
+            }
+        }
+
+        /// <summary>
+        ///     Send a BUI message to a specific player session.
+        /// </summary>
+        public bool TrySendUiMessage(EntityUid uid, Enum uiKey, BoundUserInterfaceMessage message, ICommonSession session, UserInterfaceComponent? ui = null)
+        {
+            if (!TryGetUi(uid, uiKey, out var bui, ui))
+                return false;
+
+            return TrySendUiMessage(bui, message, session);
+        }
+
+        /// <summary>
+        ///     Send a BUI message to a specific player session.
+        /// </summary>
+        public bool TrySendUiMessage(PlayerBoundUserInterface bui, BoundUserInterfaceMessage message, ICommonSession session)
+        {
+            if (!bui.SubscribedSessions.Contains(session))
+                return false;
+
+            RaiseNetworkEvent(new BoundUIWrapMessage(GetNetEntity(bui.Owner), message, bui.UiKey), session.Channel);
+            return true;
+        }
+
+        #endregion
+    }
+
+    /// <summary>
+    /// Raised by <see cref="UserInterfaceSystem"/> to check whether an interface is still accessible by its user.
+    /// </summary>
+    [ByRefEvent]
+    [PublicAPI]
+    public struct BoundUserInterfaceCheckRangeEvent
+    {
+        /// <summary>
+        /// The entity owning the UI being checked for.
+        /// </summary>
+        public readonly EntityUid Target;
+
+        /// <summary>
+        /// The UI itself.
+        /// </summary>
+        /// <returns></returns>
+        public readonly Enum UiKey;
+
+        /// <summary>
+        /// The player for which the UI is being checked.
+        /// </summary>
+        public readonly ICommonSession Player;
+
+        /// <summary>
+        /// The result of the range check.
+        /// </summary>
+        public BoundUserInterfaceRangeResult Result;
+
+        public BoundUserInterfaceCheckRangeEvent(
+            EntityUid target,
+            Enum uiKey,
+            ICommonSession player)
+        {
+            Target = target;
+            UiKey = uiKey;
+            Player = player;
+        }
+    }
+}
+
+/// <summary>
+/// Possible results for a <see cref="BoundUserInterfaceCheckRangeEvent"/>.
+/// </summary>
+public enum BoundUserInterfaceRangeResult : byte
+{
+    /// <summary>
+    /// Run built-in range check.
+    /// </summary>
+    Default,
+
+    /// <summary>
+    /// Range check passed, UI is accessible.
+    /// </summary>
+    Pass,
+
+    /// <summary>
+    /// Range check failed, UI is inaccessible.
+    /// </summary>
+    Fail
 }
